pyproject.toml
[project]
name = "law-rag-system"
version = "0.1.0"
description = "Legal RAG System with MCP Architecture"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.44.1",
    "langchain-chroma>=0.2.3",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langgraph>=0.3.21",
    "langchain-openai>=0.3.11",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=2.2.2",
    "nest-asyncio>=1.6.0",
    "chromadb>=0.4.18"
]